--- a/src-docs/webpack.config.js
+++ b/src-docs/webpack.config.js
@@ -164,6 +164,8 @@ const webpackConfig = {
       }),
     ],
   },
+
+  stats: 'minimal',
 };
 
 // Inspired by `get-port-sync`, but propogates options