@constancecchen 2022-05-26
Per https://stackoverflow.com/a/37664373/4294462 this should also speed up our local dev and CI compile times